refactor(helpers): better naming, clearer docs, no 'raw_pattern' key
This commit refactors the helpers submodule. The largest change is the removal
of the 'raw_pattern' key of the RegexRules object. The reason being is that
there is no plans to support other keys other than that, so it would be much
simpler and straightforward to flatten the nested dictionary.

Documentation is also added in this commit to aid in navigating the submodule.
Several argument names were also changed or made more explicit to reduce any
ambiguity as to what they are used for.