refactor: make function private, add docs, use clearer names
This commit performs a few refactors. The first is the privatisation of the
'failed_response' function that is never meant for users to access in the first
place. The second is adding comprehensive docstrings to all the functions of
this module. Lastly, several arguments are renamed in favour of clarity.