feat!: remove subparsers, make parser basic
The breaking change in this commit is that it removes both previously defined
subparsers and replaces them with a single parser at the root level. The reason
for the change is that for the purpose of this project, there is no need for the
steps to be separated and to have different parsers to handle the different
steps. A single parser to take in information of the container image and the
base tag to compare it against is more than enough.