feat!: rename ciu to image_version_checker
This commit changes the name of the script from 'ciu' to 'image_version_checker'
to properly reflect the purpose of the script. Indeed, there is no updating of a
compose file in the script, nor are there plans for it. This project (and hence
this script) is at the most only to check for and retrieve the most up-to-date
semantic version tagged to a container image. The name of the parser is also
updated to reflect this change.

Besides these changes, documentation is added to the parser module to describe
how it should be used.